Completed the Search Bar
deleted it from Appjs to try and avoid merge conflict
